Design Lab: Finished impleplementing ioctl and added user level program to conduct ioctl. User program was also added to makefile.

+++ crashtest.c
@@ -10,8 +10,8 @@ int main(int argc, char **argv)
 {
 	char *file_name = "./test/ospfs_crash";
 	int fd;
-	unsigned long nwrites;
-	unsigned long ioctl_retval;
+	int nwrites;
+	int ioctl_retval;
 	
 	if(argc == 2)
 	{
